Move Resolve impl for Filter to render::filters
